Fix cron logic, general cleanup

+++ lego.sh
@@ -3,23 +3,16 @@
 set -e
 
 BASEDIR=$(dirname "$0")
+. "${BASEDIR}"/lego.env
 
 if [ "${BASEDIR}" = '.' ]; then
-	echo 'This script should be run via an absolute path.'
+	echo 'This script should be run via an absolute path'
 	exit 1
 fi
 
-. "${BASEDIR}"/lego.env
-
-CRON_FILE='/etc/cron.daily/lego'
-# You might have to change this next line depending on what DNS provider you use
-PODMAN_ENV="-e CLOUDFLARE_API_KEY=${CLOUDFLARE_API_KEY} -e CLOUDFLARE_EMAIL=${CLOUDFLARE_EMAIL}"
-PODMAN_CMD="podman run -it --rm --name=lego --network=host ${PODMAN_ENV} -v ${SSL_PATH}/lego/:/var/lib/lego/ hectormolinero/lego"
-LEGO_ARGS="--dns ${DNS_PROVIDER} --domains ${CERT_HOST} --email ${CERT_EMAIL}"
-
-if [ ! -f "${CRON_FILE}" ]; then
-	echo "sh ${SSL_PATH}/lego.sh renew" >${CRON_FILE}
-	chmod 700 ${CRON_FILE}
+if [ "${BASEDIR}" != "${SSL_PATH}" ]; then
+	echo "The directory in which this script resides (${BASEDIR}) does not match what SSL_PATH is set to (${SSL_PATH})"
+	exit 1
 fi
 
 deploy_cert() {
@@ -27,12 +20,31 @@ deploy_cert() {
 	KEY="${SSL_PATH}/lego/certificates/${CERT_HOST}.key"
 	CERT_PATH='/mnt/data/unifi-os/unifi-core/config'
 
-	cp -f "${CERT}" ${CERT_PATH}/unifi-core.crt
-	cp -f "${KEY}" ${CERT_PATH}/unifi-core.key
-	chmod 644 ${CERT_PATH}/unifi-core.*
+	if [ "$(find -L "${SSL_PATH}"/lego -type f -name "${CERT_HOST}".crt -mmin -5)" ]; then
+		echo 'New certificate was generated, time to deploy it'
+		cp -f "${CERT}" ${CERT_PATH}/unifi-core.crt
+		cp -f "${KEY}" ${CERT_PATH}/unifi-core.key
+		chmod 644 ${CERT_PATH}/unifi-core.*
 
-	unifi-os restart
+		unifi-os restart
+	else
+		echo 'No new certificate was found, exiting without restart'
+	fi
 }
+
+# You might have to change this next line depending on what DNS provider you use
+# Using --env-file with podman doesn't seem to export the Cloudflare variables
+# so that lego sees them.
+PODMAN_ENV="-e CLOUDFLARE_API_KEY=${CLOUDFLARE_API_KEY} -e CLOUDFLARE_EMAIL=${CLOUDFLARE_EMAIL}"
+PODMAN_CMD="podman run -it --rm --name=lego --network=host ${PODMAN_ENV} -v ${SSL_PATH}/lego/:/var/lib/lego/ hectormolinero/lego"
+LEGO_ARGS="--dns ${DNS_PROVIDER} --domains ${CERT_HOST} --email ${CERT_EMAIL}"
+
+CRON_FILE='/etc/cron.d/lego'
+if [ ! -f "${CRON_FILE}" ]; then
+	echo "0 3 * * * sh ${SSL_PATH}/lego.sh renew" > ${CRON_FILE}
+	chmod 644 ${CRON_FILE}
+	/etc/init.d/crond reload ${CRON_FILE}
+fi
 
 case $1 in
 initial)
@@ -42,19 +54,10 @@ initial)
 	fi
 
 	echo 'Attempting initial certificate generation'
-	${PODMAN_CMD} ${LEGO_ARGS} --accept-tos --key-type rsa2048 run
-
-	if [ $? -ne 1 ]; then
-		echo 'Certificate generation was successful, deploying certificate'
-		deploy_cert
-	fi
+	${PODMAN_CMD} ${LEGO_ARGS} --accept-tos --key-type rsa2048 run && deploy_cert
 	;;
 renew)
-	${PODMAN_CMD} ${LEGO_ARGS} renew --days 80
-	# We can't use the `--renewal-hook` easily with podman
-	if [ "$(find -L "${SSL_PATH}"/lego -type f -name "${CERT_HOST}".crt -mtime -1)" ]; then
-		echo 'Certificate renewal was successful, deploying certificate'
-		deploy_cert
-	fi
+	echo 'Attempting certificate renewal'
+	${PODMAN_CMD} ${LEGO_ARGS} renew --days 80 && deploy_cert
 	;;
 esac
